requirements.txt
Flask
gevent
gunicorn
keras
numpy
h5py
pillow

